requirements.txt
django
djangorestframework
django-cors-headers
PyJWT
tensorflow
tensorflow_recommenders
pandas
